Use [~/.config/mybash]. Externalize [900.user].
The scripts in [900.user] are user preference settings that are
liable to change but which add nothing to the underlying borg
functionality.

The dothandler will (via [./first]) consult user configuration
settings by sourcing [~/.config/mysource/activate]. This will
set various [$MYBORG_CONFIG_*] environment variables for the
proxy dot.* scripts to use.

In particular: the environment variable [$MYBORG_CONFIG_MODULES],
if defined, will be assumed to point to a directory containing
user modules to be sourced by borg.

It is the responsability of the user to maintain the script
[~/.config/mysource/activate] and to have any appropriate
[$MYBORG_CONFIG_MODULES] populated.